+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Removed
+- Removed `spicyweb\embeddedassets\Service::getCachedAssetPath()`
+
 ## 2.9.0 - 2021-11-03
 
 ### Added